Isolate format conversion from engine facade
convert now lives in a dedicated module while mcp_video.engine.convert remains available for public callers. The extraction preserves format-specific behavior, progress callbacks, two-pass encoding, GIF palette generation, and thumbnail reporting.

Constraint: Keep timeline/composite code untouched; convert has broad codec/format behavior and needed its own review lane.

Rejected: Change quality preset plumbing or GIF settings here | this PR is an extraction, not an encoding-policy redesign.

Confidence: high

Scope-risk: broad

Directive: Preserve progress callback and two-pass behavior when modifying conversion internals.

Tested: ruff check mcp_video/engine.py mcp_video/engine_convert.py; ruff format --check mcp_video/engine.py mcp_video/engine_convert.py; python3 -m pytest tests/test_engine.py -k 'convert or gif or progress' tests/test_engine_advanced.py -k 'convert or two_pass or export_video' -q --tb=short; python3 -m pytest tests/test_cli.py -k 'convert' tests/test_server.py -k 'convert' tests/test_client.py -k 'convert' tests/test_red_team.py -k 'convert' -q --tb=short; python3 -m pytest tests/test_engine.py tests/test_e2e.py tests/test_server.py -q --tb=short

Not-tested: Full slow/real-media suite.
